requirements.txt
streamlit==1.39.0
groq
python-dotenv
gTTS
